Fix a race condition
The progress page would sometimes be sent back, before the content
of the multimedia tag. Fix this by adding a GET parameter to
disable the progress page. Note that this behavior is problably
a bug in the browser, which does not populate the ACCEPT header
field with a multimedia MIME type.